Add shared library and Ruby example
