+++ Cargo.toml
@@ -17,6 +17,7 @@ lto = "thin"
 [dependencies]
 wasm-bindgen = { version = "0.2", features = ["serde-serialize"] }
 
+libc = "0.2.114"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 js-sys = "0.3"
